Switch win32 implementation to new design
