+++ src/os/windows.rs
@@ -18,7 +18,7 @@ pub trait WindowExt {
 impl WindowExt for Window {
     #[inline]
     fn get_hwnd(&self) -> *mut libc::c_void {
-        self.window.platform_window()
+        self.window.hwnd() as *mut _
     }
 }
